Mark tests that sleep as @medium (#94)
* Fix expiration tests

Setting a +1 second expiration is very unreliable.
A +1 second expiration doesn't mean the value will expire
"in a second", but "next second". If we're already at
950ms within our second, then by the time "set" and the
follow-up "get" go out, we may already be within the
next second. As a result, the item will already have been
expired and we would no longer be able to find the value.

Better to be safe and set +2 seconds (which will give us
anywhere between 1 and 2 seconds)

Meanwhile also reduced the amount of sleeps: we could
manage with only 1 per function.

* Mark tests that sleep with `@medium`

* Trigger travis

Co-authored-by: Matthias Mullie <git@mullie.eu>
Co-authored-by: Tobias Nyholm <tobias.nyholm@gmail.com>